move suggested users query down to tab
